third_party/blink/web_tests/external/wpt/custom-elements/registries/per-global-expected.txt

This is a testharness.js-based test.
[FAIL] Navigating from the initial about:blank must not replace window.customElements
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
